+++ packages/xy/src/scatter/defaults.ts
@@ -1,0 +1,10 @@
+import { X, Y } from '@chsk/core'
+import { ScatterInteractiveDataItem } from './types'
+
+export const defaultScatterTooltipFormat = (data: ScatterInteractiveDataItem) => {
+    if (data.point) {
+        return String(data.index) + ' (' + data.point[X] + ', ' + data.point[Y] + ')'
+    } else {
+        return String(data.index)
+    }
+}