+++ packages/xy/stories/scatter/decorators.tsx
@@ -1,6 +1,12 @@
 import { ReactNode } from 'react'
-import { Axis, Chart, Legend, Circle, SymbolProps, Square } from '@chsk/core'
-import { Scatter, ScatterCurve, ScatterPoints, ScatterInteractiveDataItem } from '../../src'
+import { Axis, Chart, Legend, Circle, SymbolProps, Square, Tooltip } from '@chsk/core'
+import {
+    Scatter,
+    ScatterCurve,
+    ScatterPoints,
+    ScatterInteractiveDataItem,
+    ScatterCrosshair,
+} from '../../src'
 import dataPolynomials from './dataPolynomials.json'
 import { generateScatterSeries } from './generators'
 
@@ -57,20 +63,22 @@ export const ChartWithLegendSpaceDecorator = (Story: () => ReactNode) => (
     </Chart>
 )
 
+const round3dp = (x: number) => Math.round(x * 1000) / 1000
+
 export const dataRegression = [
     generateScatterSeries(
         'A',
         Array(16)
             .fill(0)
-            .map((v, i) => i / 2 + Math.random()),
-        x => 5 + 1.1 * x + Math.random() * 5
+            .map((v, i) => round3dp(i / 2 + Math.random())),
+        x => round3dp(5 + 1.1 * x + Math.random() * 5)
     ),
     generateScatterSeries(
         'B',
         Array(16)
             .fill(0)
-            .map((v, i) => i / 2 + Math.random()),
-        x => 3 + 0.3 * x + Math.random() * 4
+            .map((v, i) => round3dp(i / 2 + Math.random())),
+        x => round3dp(3 + 0.3 * x + Math.random() * 4)
     ),
 ]
 
@@ -90,10 +98,10 @@ export const ChartForRegressionDecorator = (Story: () => ReactNode) => (
                 domain: [0, 'auto'],
             }}
         >
-            <Axis variant={'bottom'} label={'x (a.u.)'} />
-            <Axis variant={'left'} label={'y (a.u.)'} />
             <ScatterPoints />
             {Story()}
+            <Axis variant={'bottom'} label={'x (a.u.)'} />
+            <Axis variant={'left'} label={'y (a.u.)'} />
             <Legend
                 position={[220, 160]}
                 anchor={[0, 0.5]}
@@ -132,3 +140,41 @@ export const DoubleSquare = (props: SymbolProps) => {
         </>
     )
 }
+
+export const ChartWithTooltipDecorator = () => (
+    <Chart size={[400, 300]} padding={[40, 140, 60, 60]} style={{ display: 'inline-block' }}>
+        <Scatter
+            data={dataRegression}
+            x={'x'}
+            y={'y'}
+            valueSize={4}
+            scaleX={{
+                variant: 'linear',
+                domain: [0, 'auto'],
+            }}
+            scaleY={{
+                variant: 'linear',
+                domain: [0, 'auto'],
+            }}
+        >
+            <Axis variant={'bottom'} label={'x (a.u.)'} />
+            <Axis variant={'left'} label={'y (a.u.)'} />
+            <ScatterPoints />
+            <ScatterCrosshair style={{ stroke: '#444', strokeWidth: 1, strokeDasharray: 6 }}>
+                <Tooltip
+                    position={[0, -10]}
+                    anchor={[0.5, 1]}
+                    itemSize={[160, 32]}
+                    itemPadding={[8, 8, 8, 8]}
+                    style={{ stroke: '#222222', strokeWidth: 1 }}
+                />
+            </ScatterCrosshair>
+            <Legend
+                position={[220, 160]}
+                anchor={[0, 0.5]}
+                positionUnits={'absolute'}
+                symbol={Circle}
+            />
+        </Scatter>
+    </Chart>
+)